Add number of quotes on stats page. Close #10
