+++ gestion/urls.py
@@ -52,7 +52,6 @@ urlpatterns = [
     path('categoryProfile/<int:pk>', views.categoryProfile, name="categoryProfile"),
     path('categoriesList', views.categoriesList, name="categoriesList"),
     path('categories-autocomplete', views.CategoriesAutocomplete.as_view(), name="categories-autocomplete"),
-    path('stats', views.stats, name="stats"),
     path('divide', views.divide, name="divide"),
     path('gen_invoice', views.gen_invoice, name="gen_invoice"),
     path('compute-price', views.compute_price_view, name="compute-price"),
