+++ templates/stats.html
@@ -64,6 +64,10 @@
 					<td>{{superusers.count}}</td>
 				</tr>
 				<tr>
+					<td>Nombre de citations</td>
+					<td>{{nb_quotes}}</td>
+				</tr>
+				<tr>
 					<td>Nombre 8</td>
 					<td>8</td>
 				</tr>
